+++ src/main.h
@@ -287,7 +287,7 @@ class MainApp : public wxApp
         bool                m_psk_enable;
         wxString            m_psk_callsign;
         wxString            m_psk_grid_square;
-        int                 m_psk_freq;
+        uint64_t            m_psk_freq;
 
         PskReporter*            m_pskReporter;
         std::string         m_pskPendingCallsign;
